+++ lib/rfid_positioning/tag_angles.h
@@ -1,6 +1,7 @@
 #include <map>
+#include <cstdint>
 
-std::map<int, int> tags{
+std::map<uint64_t, int> tags{
     {2462619523, 0},
     {37038852, 1},
     {946104749, 2}
